requirements.txt
Flask==1.1.2
pandas==1.1.4
numpy==1.19.4
python-decouple==3.4
gunicorn==19.9.0
